+++ app/ui/src/components/FlightRecord/FlightRecord.jsx
@@ -64,6 +64,8 @@ export const FlightRecord = () => {
 
   const gridSize = useMemo(() => ({ xs: 12, sm: 12, md: 6, lg: 6, xl: 6 }), []);
 
+  const options = flight?.track ? { noRoutes: true, noTracks: false } : { noRoutes: false, noTracks: true };
+
   return (
     <>
       {isLoading && <LinearProgress />}
@@ -75,7 +77,7 @@ export const FlightRecord = () => {
         </Grid>
 
         <Grid size={gridSize}>
-          <FlightMap data={mapData} />
+          <FlightMap data={mapData} options={options} />
         </Grid>
       </Grid>
     </>