+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-- Update: Add `numeric` property for Time and Landing fields to show a numeric keyboard on mobile devices for better UX
+- Update: Add `numeric` property for Time and Landing fields to show a numeric keyboard on mobile devices for better UX @danielkappelle
 
 ## [3.6.0] - 14.08.2025